requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow-cpu==2.19.0
pillow==10.1.0
python-dotenv==1.0.0 